feat: fixed bugs when caps and shift were pressed, add new styles

+++ script.js
@@ -3,11 +3,11 @@ const textarea = document.createElement('textarea');
 textarea.setAttribute('autofocus', 'autofocus');
 const keysContainer = document.createElement('div');
 keysContainer.className = 'keys-container';
+let caps = false;
 
 body.appendChild(textarea);
 body.appendChild(keysContainer);
-
-body.insertAdjacentHTML('beforeend', '<p class="instruction">Переключение языка ввода виртуальной клавиатуры происходит одновременным нажатием клавиш LeftShift и LeftCtrl на физической клавиатуре. Клавиатура была сделана в системе Windows 10.</p>');
+body.insertAdjacentHTML('beforeend', '<p class="instruction">Переключение языка ввода виртуальной клавиатуры происходит одновременным нажатием клавиш LeftShift и LeftCtrl на физической клавиатуре. Клавиатура была сделана в системе Windows</p>');
 
 function disable() { document.onkeydown = () => false; } disable();
 
@@ -16,36 +16,32 @@ const engListDef = [
   ['Tab', 'q', 'w', 'e', 'r', 't', 'y', 'u', 'i', 'o', 'p', '[', ']', 'Del'],
   ['Caps</br>Lock', 'a', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', "'", '\\', 'Enter'],
   ['Shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/', '&#9652', 'Shift'],
-  ['Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656'],
+  ['Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656', '&#128125'],
 ];
-
 const engListUp = [
   '~', '!', '@', '#', '$', '%', '^', '&', '*', '(', ')', '_', '+', 'Backspace',
   'Tab', 'Q', 'W', 'E', 'R', 'T', 'Y', 'U', 'I', 'O', 'P', '{', '}', 'Del',
   'Caps</br>Lock', 'A', 'S', 'D', 'F', 'G', 'H', 'J', 'K', 'L', ':', '"', '|', 'Enter',
   'Shift', 'Z', 'X', 'C', 'V', 'B', 'N', 'M', '<', '>', '?', '&#9652', 'Shift',
-  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656',
+  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656', '&#128125',
 ];
-
 const rusListDef = [
   'ё', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'Backspace',
   'Tab', 'й', 'ц', 'у', 'к', 'е', 'н', 'г', 'ш', 'щ', 'з', 'х', 'ъ', 'Del',
   'Caps</br>Lock', 'ф', 'ы', 'в', 'а', 'п', 'р', 'о', 'л', 'д', 'ж', 'э', '\\', 'Enter',
   'Shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'б', 'ю', '.', '&#9652', 'Shift',
-  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656',
+  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656', '&#128125',
 ];
-
 const rusListUp = [
   'Ё', '!', '"', '№', ';', '%', ':', '?', '*', '(', ')', '_', '+', 'Backspace',
   'Tab', 'Й', 'Ц', 'У', 'К', 'Е', 'Н', 'Г', 'Ш', 'Щ', 'З', 'Х', 'Ъ', 'Del',
   'Caps</br>Lock', 'Ф', 'Ы', 'В', 'А', 'П', 'Р', 'О', 'Л', 'Д', 'Ж', 'Э', '/', 'Enter',
   'Shift', 'Я', 'Ч', 'С', 'М', 'И', 'Т', 'Ь', 'Б', 'Ю', ',', '&#9652', 'Shift',
-  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656',
+  'Ctrl', 'Win', 'Alt', ' ', 'Alt', 'Ctrl', '&#9666', '&#9662', '&#9656', '&#128125',
 ];
-
 const keysList = ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace',
   'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Delete',
-  'Caps</br>Lock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Backslash', 'Enter',
+  'CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Backslash', 'Enter',
   'ShiftLeft', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp', 'ShiftRight',
   'ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ControlRight', 'ArrowLeft', 'ArrowDown', 'ArrowRight',
 ];
@@ -77,11 +73,7 @@ let lang;
 let languageStorage;
 let language = JSON.parse(localStorage.getItem(lang));
 
-if (!language) {
-  valuesKeys(oneDimensionalList);
-} else {
-  valuesKeys(language);
-}
+if (!language) { valuesKeys(oneDimensionalList); } else { valuesKeys(language); }
 
 const backspace = allKeys[13];
 const tab = allKeys[14];
@@ -100,6 +92,7 @@ const ctrlRight = allKeys[60];
 const arrLeft = allKeys[61];
 const arrRight = allKeys[63];
 const arrDown = allKeys[62];
+const smile = allKeys[64];
 
 backspace.classList.add('backspace');
 backspace.classList.add('functional');
@@ -138,6 +131,7 @@ arrRight.classList.add('functional');
 arrDown.classList.add('arrow');
 arrDown.classList.add('arrowDown');
 arrDown.classList.add('functional');
+smile.classList.add('smile');
 
 document.addEventListener('keydown', (event) => {
   let cursorPosition = textarea.selectionStart;
@@ -206,11 +200,31 @@ document.addEventListener('keydown', (event) => {
   }
 
   if (event.code === 'ShiftLeft' || event.code === 'ShiftRight') {
-    if (allKeys[0].innerHTML === '`') {
-      valuesKeys(engListUp);
+    if (caps === true) {
+      if (allKeys[0].innerHTML === '`') {
+        valuesKeys(engListUp);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+          }
+        }
+      }
+      if (allKeys[0].innerHTML === 'ё') {
+        valuesKeys(rusListUp);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+          }
+        }
+      }
     }
-    if (allKeys[0].innerHTML === 'ё') {
-      valuesKeys(rusListUp);
+    if (caps === false) {
+      if (allKeys[0].innerHTML === '`') {
+        valuesKeys(engListUp);
+      }
+      if (allKeys[0].innerHTML === 'ё') {
+        valuesKeys(rusListUp);
+      }
     }
   }
 });
@@ -224,38 +238,39 @@ document.addEventListener('keyup', (event) => {
   }
 
   if (event.code === 'ShiftLeft' || event.code === 'ShiftRight') {
-    if (allKeys[0].innerHTML === '~') {
-      valuesKeys(oneDimensionalList);
+    if (caps === true) {
+      if (allKeys[0].innerHTML === '~') {
+        valuesKeys(oneDimensionalList);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+          }
+        }
+      }
+      if (allKeys[0].innerHTML === 'Ё') {
+        valuesKeys(rusListDef);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+          }
+        }
+      }
     }
-    if (allKeys[0].innerHTML === 'Ё') {
-      valuesKeys(rusListDef);
+    if (caps === false) {
+      if (allKeys[0].innerHTML === '~') {
+        valuesKeys(oneDimensionalList);
+      }
+      if (allKeys[0].innerHTML === 'Ё') {
+        valuesKeys(rusListDef);
+      }
     }
   }
 });
 
-function switchLanguage(func, ...codes) {
-  const pressed = new Set();
-
-  document.addEventListener('keydown', (event) => {
-    pressed.add(event.code);
-
-    for (const code of codes) {
-      if (!pressed.has(code)) {
-        return;
-      }
-    }
-    pressed.clear();
-
-    func();
-  });
-
-  document.addEventListener('keyup', (event) => {
-    pressed.delete(event.code);
-  });
-}
-
-switchLanguage(
-  () => {
+document.addEventListener('keydown', (e) => {
+  if (e.ctrlKey && e.shiftKey) {
+    caps = false;
+    allKeys[28].classList.remove('keystroke');
     if (allKeys[0].innerHTML === '~') {
       localStorage.clear();
       languageStorage = rusListDef;
@@ -271,24 +286,27 @@ switchLanguage(
       language = JSON.parse(localStorage.getItem(lang));
       valuesKeys(oneDimensionalList);
     }
-  },
-  'ShiftLeft',
-  'ControlLeft',
-);
+  }
+});
 
 document.addEventListener('keyup', (event) => {
-  if (event.getModifierState('CapsLock')) {
-    for (let i = 0; i < allKeys.length; i += 1) {
-      if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
-        allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+  if (event.code === 'CapsLock') {
+    if (caps === false) {
+      allKeys[28].classList.add('keystroke');
+      for (let i = 0; i < allKeys.length; i += 1) {
+        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+          allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+        }
       }
-    }
-  }
-  if (!event.getModifierState('CapsLock')) {
-    for (let i = 0; i < allKeys.length; i += 1) {
-      if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
-        allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+      caps = true;
+    } else {
+      allKeys[28].classList.remove('keystroke');
+      for (let i = 0; i < allKeys.length; i += 1) {
+        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+          allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+        }
       }
+      caps = false;
     }
   }
 });
@@ -311,6 +329,26 @@ function virtualInput(event) {
   let cursorPositionEnd = textarea.selectionEnd;
   const beforeText = textarea.value.slice(0, cursorPosition);
   const afterText = textarea.value.slice(cursorPosition);
+
+  if (event.target.classList.contains('capsLock')) {
+    if (caps === false) {
+      allKeys[28].classList.add('keystroke');
+      for (let i = 0; i < allKeys.length; i += 1) {
+        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+          allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+        }
+      }
+      caps = true;
+    } else {
+      allKeys[28].classList.remove('keystroke');
+      for (let i = 0; i < allKeys.length; i += 1) {
+        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+          allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+        }
+      }
+      caps = false;
+    }
+  }
 
   if (event.target.classList.contains('functional') === false
     && event.target.classList.contains('key')) {
@@ -374,33 +412,38 @@ function virtualInput(event) {
 keysContainer.addEventListener('click', virtualInput);
 
 function virtualPress(event) {
-  if (event.target.classList.contains('capsLock')) {
-    if (event.target.classList.contains('keystroke')) {
-      event.target.classList.remove('keystroke');
-      for (let i = 0; i < allKeys.length; i += 1) {
-        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
-          allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
-        }
-      }
-    } else {
-      event.target.classList.add('keystroke');
-      for (let i = 0; i < allKeys.length; i += 1) {
-        if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
-          allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
-        }
-      }
-    }
-  }
-
-  if (event.target.classList.contains('shiftLeft') || event.target.classList.contains('shiftRight')) {
+  if (event.target.classList.contains('shiftLeft')
+    || event.target.classList.contains('shiftRight')) {
     event.target.classList.add('keystroke');
-    if (allKeys[0].innerHTML === '`') {
-      valuesKeys(engListUp);
-    } else {
-      valuesKeys(rusListUp);
+    if (caps === true) {
+      if (allKeys[0].innerHTML === '`') {
+        valuesKeys(engListUp);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+          }
+        }
+      }
+      if (allKeys[0].innerHTML === 'ё') {
+        valuesKeys(rusListUp);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toLowerCase();
+          }
+        }
+      }
+    }
+    if (caps === false) {
+      if (allKeys[0].innerHTML === '`') {
+        valuesKeys(engListUp);
+      }
+      if (allKeys[0].innerHTML === 'ё') {
+        valuesKeys(rusListUp);
+      }
     }
   }
-  if (event.target.tagName === 'LI') {
+  if (event.target.tagName === 'LI'
+    && !event.target.classList.contains('capsLock')) {
     event.target.classList.add('keystroke');
   }
 }
@@ -408,15 +451,38 @@ function virtualPress(event) {
 keysContainer.addEventListener('mousedown', virtualPress);
 
 function virtualRelease(event) {
-  if (event.target.classList.contains('shiftLeft') || event.target.classList.contains('shiftRight')) {
+  if (event.target.classList.contains('shiftLeft')
+    || event.target.classList.contains('shiftRight')) {
     event.target.classList.remove('keystroke');
-    if (allKeys[0].innerHTML === '~') {
-      valuesKeys(oneDimensionalList);
-    } else {
-      valuesKeys(rusListDef);
+    if (caps === true) {
+      if (allKeys[0].innerHTML === '~') {
+        valuesKeys(oneDimensionalList);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+          }
+        }
+      }
+      if (allKeys[0].innerHTML === 'Ё') {
+        valuesKeys(rusListDef);
+        for (let i = 0; i < allKeys.length; i += 1) {
+          if (/^[a-zа-я]$/i.test(allKeys[i].innerHTML)) {
+            allKeys[i].innerHTML = allKeys[i].innerHTML.toUpperCase();
+          }
+        }
+      }
+    }
+    if (caps === false) {
+      if (allKeys[0].innerHTML === '~') {
+        valuesKeys(oneDimensionalList);
+      }
+      if (allKeys[0].innerHTML === 'Ё') {
+        valuesKeys(rusListDef);
+      }
     }
   }
-  if (event.target.tagName === 'LI') {
+  if (event.target.tagName === 'LI'
+    && !event.target.classList.contains('capsLock')) {
     event.target.classList.remove('keystroke');
   }
 }
